requirements.txt
django==5.2
djangorestframework==3.16.0
celery==5.5.2
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-jalali-date==2.0.0
django-silk==5.3.2
djangorestframework-simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
flower==2.0.1
gunicorn==23.0.0
locust==2.36.2
model-bakery==1.20.4
pillow==11.2.1
pytest-django==4.11.1
pytest-watch==4.2.0
pytest==8.3.5
redis==6.0.0
